extract test utils
